feat: update test suit for journal entries
